Backend Completed

+++ src/controllers/blog_controllers.py
@@ -1,14 +1,15 @@
 from fastapi import status, HTTPException
 from sqlalchemy.orm import Session
 from src.models.blog_models import BlogModel, BlogSchema, BlogCategory
+from src.models.user_models import UserModel
 import logging
 from typing import Optional
 
 logger = logging.getLogger(__name__)
 
-def get_all_blogs(db:Session, category:Optional[BlogCategory], search:Optional[str]):
+def get_all_blogs(db:Session, category:Optional[BlogCategory], search:Optional[str], user:UserModel):
     try:
-        query = db.query(BlogModel)
+        query = db.query(BlogModel).filter(BlogModel.user_id == user.id)
         
         if category:
             query = query.filter(BlogModel.category == category)
@@ -26,7 +27,7 @@ def get_all_blogs(db:Session, category:Optional[BlogCategory], search:Optional[s
             detail=f'Failed to get all blogs'
         )
 
-def get_blog(id:int, db:Session):
+def get_blog(id:int, db:Session, user:UserModel):
     try:
         blog = db.query(BlogModel).get(id)
 
@@ -35,6 +36,13 @@ def get_blog(id:int, db:Session):
             raise HTTPException(
                 status_code=status.HTTP_404_NOT_FOUND,
                 detail='Blog not found'
+            )
+        
+        if blog.user_id != user.id:
+            logger.error('You are not authorized to access this blog')
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail='You are not authorized to access this blog'
             )
         
         return blog
@@ -46,12 +54,13 @@ def get_blog(id:int, db:Session):
             detail='Failed to get the blog'
         )
 
-def create_blog(request:BlogSchema, db:Session):
+def create_blog(request:BlogSchema, db:Session, user:UserModel):
     try:
         new_blog = BlogModel(
             title = request.title,
             content = request.content,
-            category = request.category
+            category = request.category,
+            user_id = user.id
         )
 
         db.add(new_blog)
@@ -67,7 +76,7 @@ def create_blog(request:BlogSchema, db:Session):
             detail='Failed to create the blog'
         )
 
-def update_blog(id:int, request:BlogSchema, db:Session):
+def update_blog(id:int, request:BlogSchema, db:Session, user:UserModel):
     try:
         blog = db.query(BlogModel).get(id)
 
@@ -76,6 +85,13 @@ def update_blog(id:int, request:BlogSchema, db:Session):
             raise HTTPException(
                 status_code=status.HTTP_404_NOT_FOUND,
                 detail='Blog not found'
+            )
+        
+        if blog.user_id != user.id:
+            logger.error('You are not authorized to update this blog')
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail='You are not authorized to update this blog'
             )
         
         blog.title = request.title
@@ -94,7 +110,7 @@ def update_blog(id:int, request:BlogSchema, db:Session):
             detail='Failed to update the blog'
         )
 
-def delete_blog(id:int, db:Session):
+def delete_blog(id:int, db:Session, user:UserModel):
     try:
         blog = db.query(BlogModel).get(id)
 
@@ -103,6 +119,13 @@ def delete_blog(id:int, db:Session):
             raise HTTPException(
                 status_code=status.HTTP_404_NOT_FOUND,
                 detail='Blog not found'
+            )
+        
+        if blog.user_id != user.id:
+            logger.error('You are not authorized to delete this blog')
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail='You are not authorized to delete this blog'
             )
         
         db.delete(blog)
